Move Mesh.to_dict/from_dict to skfem.io.json
